[fix] No wordwrap no longer cause cowthink to read from stdin
The command `echo foo | cowthink "" bar` will say foo, not bar.
The emptyness of the first argument decides which source of input cowsay will use.
By default pokemonsay will pass in a "$word_wrap" which by default evaluates to "".
This causes cowsay to read from stdin, even if a message is provided via program arguments.

This change uses bash arrays to properly expand the "$word_wrap" into arguments.
This neccesitated a change of the shebang, but the pokemonsay was not posix-shell compliant anyway due to the "^" on line 50.